modify bubble sort
